move integration tests into its own package
