+++ test/main_test.go
@@ -1,11 +1,7 @@
-package devastator_test
-
-// todo: should package be different to make this into a true integration test from a client perspective?
-
-// this is the integration test package from a real client perspective.
+// Package test is an integration testing package for testing the server from a mobile client's perspective.
+package test
 
 import (
-	"fmt"
 	"net"
 	"os"
 	"sync"
@@ -27,126 +23,6 @@ var (
 	clientCertBytes = []byte(clientCert)
 	clientKeyBytes  = []byte(clientKey)
 )
-
-func TestClientDisconnect(t *testing.T) {
-	// todo: we need to verify that events occur in the order that we want them (either via event hooks or log analysis)
-	// this seems like a listener test than a integration test from a client perspective
-	s := getServer(t)
-	c := getClientConnWithClientCert(t)
-	if err := c.Close(); err != nil {
-		t.Fatal("Failed to close the client connection:", err)
-	}
-	if err := s.Stop(); err != nil {
-		t.Fatal("Failed to stop the server:", err)
-	}
-	wg.Wait()
-}
-
-func TestClientClose(t *testing.T) {
-	// t.Fatal("Client method.close request was not handled properly")
-}
-
-func TestSendClose(t *testing.T) {
-	// t.Fatal("Server method.close request was not handled properly")
-}
-
-func TestServerDisconnect(t *testing.T) {
-	// t.Fatal("Server disconnect was not handled gracefully")
-}
-
-func TestServerClose(t *testing.T) {
-	s := getServer(t)
-	c := getClientConnWithClientCert(t)
-	if err := s.Stop(); err != nil {
-		t.Fatal("Failed to stop the server:", err)
-	}
-	if err := c.Close(); err != nil {
-		t.Fatal("Failed to close the client connection:", err)
-	}
-	wg.Wait()
-
-	// test what happens when there are outstanding connections and/or requests that are being handled
-	// destroying queues and other stuff during Close() might cause existing request handles to malfunction
-}
-
-func TestAuth(t *testing.T) {
-	// t.Fatal("Unauthorized clients cannot call any function other than method.auth.")
-}
-
-func TestGoogleAuth(t *testing.T) {
-	// t.Fatal("Google+ first sign-in (registration) failed with valid credentials")
-	// t.Fatal("Google+ second sign-in (regular) failed with valid credentials")
-	// t.Fatal("Google+ sign-in passed with invalid credentials")
-	// t.Fatal("Authentication was not ACKed")
-}
-
-func TestClientCertAuth(t *testing.T) {
-	// t.Fatal("Authentication failed with a valid client certificate")
-	// t.Fatal("Authenticated with invalid/expired client certificate")
-	// t.Fatal("Authentication was not ACKed")
-}
-
-func TestTokenAuth(t *testing.T) {
-	// t.Fatal("Authentication failed with a valid token")
-	// t.Fatal("Authenticated with invalid/expired token")
-	// t.Fatal("Authentication was not ACKed")
-}
-
-func BenchmarkAuth(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		fmt.Sprintf("hello")
-	}
-}
-
-func BenchmarkClientCertAuth(b *testing.B) {
-	// for various certificate key sizes (512....4096) and ECDSA, and with/without resumed handshake / session tickets
-	for i := 0; i < b.N; i++ {
-		fmt.Sprintf("hello")
-	}
-}
-
-func TestReceiveOfflineQueue(t *testing.T) {
-	// t.Fatal("Failed to receive queued messages after coming online")
-	// t.Fatal("Failed to send ACK for received message queue")
-}
-
-func BenchmarkQueue(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		fmt.Sprintf("hello")
-	}
-}
-
-func TestSendEcho(t *testing.T) {
-	// t.Fatal("Failed to send a message to the echo user")
-	// t.Fatal("Failed to send batch message to the echo user")
-	// t.Fatal("Failed to send large message to the echo user")
-	// t.Fatal("Did not receive ACK for a message sent")
-	// t.Fatal("Failed to receive a response from echo user")
-	// t.Fatal("Could not send an ACK for an incoming message")
-}
-
-func BenchmarkParallelThroughput(b *testing.B) {
-	// for various conn levels vs. message per second: 50:xxxx, 500:xxx, 5000:xx, ... conn/mps (hopefully!)
-	for i := 0; i < b.N; i++ {
-		fmt.Sprintf("hello")
-	}
-}
-
-func TestStop(t *testing.T) {
-	// todo: this should be a listener/queue test if we don't use any goroutines in the Server struct methods
-	// t.Fatal("Failed to stop the server: not all the goroutines were terminated properly")
-	// t.Fatal("Failed to stop the server: server did not wait for ongoing read/write operations")
-}
-
-func TestConnTimeout(t *testing.T) {
-	// t.Fatal("Send timout did not occur")
-	// t.Fatal("Wait timeout did not occur")
-	// t.Fatal("Read timeout did not occur")
-}
-
-func TestPing(t *testing.T) {
-	// t.Fatal("Pong/ACK was not sent for ping")
-}
 
 func getClientConnWithClientCert(t *testing.T) *devastator.Conn {
 	return _getClientConn(t, true)
